--- ui.py
@@ -1,5 +1,0 @@
-import streamlit as st
-
-def st_header():
-    st.header("RAGxplorer 🗺️", divider='grey')
-    st.markdown("#### Visualise which chunks are most relevant to your query.")
