+++ utils/rag.py
@@ -1,7 +1,9 @@
+"""
+Module for RAG
+"""
 import random
 import string
 import chromadb
-import umap
 import numpy as np
 import streamlit as st
 from openai import OpenAI
@@ -14,10 +16,8 @@ from langchain.text_splitter import (
 import chromadb.utils.embedding_functions as embedding_functions
 from chromadb.utils.embedding_functions import SentenceTransformerEmbeddingFunction
 from typing import (
-    Tuple, 
     List, 
     Any,
-    Union,
     TypeVar,
     )
 from typing_extensions import Protocol
@@ -161,35 +161,6 @@ def get_docs(chroma_collection: chromadb.Collection) -> List[str]:
     documents = chroma_collection.get(include=['documents'])['documents']
     return documents
 
-def set_up_umap(embeddings: np.ndarray) -> umap.UMAP:
-    """
-    Sets up and fits a UMAP transformer to the embeddings.
-    
-    Args:
-        embeddings: An array of embeddings to fit the UMAP transformer.
-    
-    Returns:
-        A fitted UMAP transformer.
-    """
-    umap_transform = umap.UMAP(random_state=0, transform_seed=0).fit(embeddings)
-    return umap_transform
-
-def get_projections(embedding: np.ndarray, umap_transform: umap.UMAP) -> Tuple[np.ndarray, np.ndarray]:
-    """
-    Projects embeddings into a two-dimensional space using UMAP.
-    
-    Args:
-        embedding: An array of embeddings to project.
-        umap_transform: A fitted UMAP transformer.
-    
-    Returns:
-        A tuple of x and y coordinates of the projected embeddings.
-    """
-    projections = _project_embeddings(embedding, umap_transform)
-    x = projections[:, 0]
-    y = projections[:, 1]
-    return x, y
-
 def get_embedding(text: str) -> np.ndarray:
     """
     Generates an embedding for the given text using a sentence transformer model.
@@ -201,22 +172,6 @@ def get_embedding(text: str) -> np.ndarray:
         An embedding of the text.
     """
     return SentenceTransformerEmbeddingFunction(model_name="all-MiniLM-L6-v2")([text])
-
-def _project_embeddings(embeddings: np.ndarray, umap_transform: umap.UMAP) -> np.ndarray:
-    """
-    Helper function to project embeddings using UMAP.
-    
-    Args:
-        embeddings: An array of embeddings to project.
-        umap_transform: A fitted UMAP transformer.
-    
-    Returns:
-        An array of projected embeddings.
-    """
-    umap_embeddings = np.empty((len(embeddings), 2))
-    for i, embedding in enumerate(embeddings):
-        umap_embeddings[i] = umap_transform.transform([embedding])
-    return umap_embeddings
 
 def _load_pdf(file: Any) -> List[str]:
     """
